+++ prisma/migrations/20251209203428_product_weight_required/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Product" ADD COLUMN     "weightInGrams" INTEGER NOT NULL DEFAULT 0;
+
+-- AlterTable
+ALTER TABLE "ProductVariant" ALTER COLUMN "weightInGrams" DROP NOT NULL;